pyproject.toml
[project]
name = "aisuiteplus"
version = "0.1.6"
description = "Uniform access layer for LLMs"
authors = [
    {name = "Vikram A", email="vikram@butterflai.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "build>=1.2.2.post1",
    "deepeval>=2.0.9",
    "hatchling>=1.27.0",
    "ipykernel>=6.29.5",
    "load-dotenv>=0.1.0",
    "loguru>=0.7.3",
    "openai>=1.58.1",
    "openpyxl>=3.1.5",
    "pypdf2>=3.0.1",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "twine>=6.0.1",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.30.1"]
aws = ["boto3>=1.34.144"]
azure = []
cohere = ["cohere>=5.12.0"]
google = ["vertexai>=1.63.0"]
groq = ["groq>=0.9.0"]
huggingface = []
mistral = ["mistralai>=1.0.3"]
ollama = []
openai = ["openai>=1.35.8"]
watsonx = ["ibm-watsonx-ai>=1.1.16"]
all = [
    "anthropic>=0.30.1",
    "boto3>=1.34.144",
    "vertexai>=1.63.0",
    "groq>=0.9.0",
    "mistralai>=1.0.3",
    "openai>=1.35.8",
    "cohere>=5.12.0",
    "ibm-watsonx-ai>=1.1.16"
]
dev = [
    "pre-commit>=3.7.1",
    "black>=24.4.2",
    "python-dotenv>=1.0.1",
    "notebook>=7.2.1",
    "ollama>=0.2.1",
    "fireworks-ai>=0.14.0",
    "chromadb>=0.5.4",
    "sentence-transformers>=3.0.1",
    "datasets>=2.20.0",
    "ipykernel>=6.29.5"
]
test = [
    "pytest>=8.2.2",
    "pytest-cov>=6.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "--cov=aisuite --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["aisuite"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
]
[project.urls]
"Homepage" = "https://github.com/Vikramardham/aisuiteplus"
"Bug Tracker" = "https://github.com/Vikramardham/aisuiteplus/issues"